Pedantic

+++ src/Timer.cc
@@ -11,5 +11,5 @@ double Timer::elapsed() const
 {
     auto currentTime = std::chrono::steady_clock::now();
     auto duration = std::chrono::duration_cast<std::chrono::microseconds>(currentTime - startTime);
-    return duration.count() / 1e6; // Second
+    return (double) duration.count() / 1e6; // Second
 }
